fix(fuse): pre-epoch timestamps — times are i64 ns in the u64 storage word (fstests generic/258)
utimensat with a negative second wrapped ('Timestamp wrapped:
18131146533'): setattr computed sec as u64 * 1e9 and the read side
divided unsigned. Times are now i64 nanoseconds carried in the existing
u64 word (two's complement — every positive value reads bit-identically,
zero on-disk format change), saturating at the i64 range at the FUSE
boundary, sign-restored with div_euclid/rem_euclid (nsec stays
non-negative, the timespec64 convention). The pending-times newest-wins
fold compares SIGNED so a backdated pre-epoch stamp cannot out-rank
echoes forever.

Green: attr_refresh 5/5 (incl. the new round-trip pin), economy/
conveyor/fold-slimming suites.

+++ src/meta_backend/kv/backend.rs
@@ -1826,10 +1826,14 @@ impl KvMetaBackend {
             return;
         }
         if let Some((pm, pc)) = self.pending_times.read_sync(&ino, |_, p| *p) {
-            if pm > v.mtime {
+            // Times are i64 ns carried in the u64 word (pre-epoch values
+            // are representable — fstests generic/258); the newest-wins
+            // fold must compare SIGNED or a backdated pre-epoch stamp
+            // would out-rank every echo forever.
+            if (pm as i64) > (v.mtime as i64) {
                 v.mtime = pm;
             }
-            if pc > v.ctime {
+            if (pc as i64) > (v.ctime as i64) {
                 v.ctime = pc;
             }
         }
